+++ CHANGELOG.md
@@ -36,6 +36,7 @@
     - Remove K Best feature selector (special case of RFE)
     - Changed Error Analysis metrics
     - Remove threat score from Multiclass Breakdown
+    - Rename Labels Are Missing exception
 
 - 0.4.1
     - Optimized CART node splitting for low variance continuous features